# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hivemind Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MarketPriceType =
  (typeof MarketPriceType)[keyof typeof MarketPriceType];

export const MarketPriceType = {
  stock: "stock",
  crypto: "crypto",
} as const;

export interface MarketPrice {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume?: number;
  marketCap?: number;
  type: MarketPriceType;
  sparkline: number[];
  updatedAt: string;
}

export interface PricePoint {
  timestamp: string;
  open?: number;
  high?: number;
  low?: number;
  close: number;
  volume?: number;
}

export interface MarketHistory {
  symbol: string;
  data: PricePoint[];
}

export type PredictionDirection =
  (typeof PredictionDirection)[keyof typeof PredictionDirection];

export const PredictionDirection = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export type PredictionOutcome =
  | (typeof PredictionOutcome)[keyof typeof PredictionOutcome]
  | null;

export const PredictionOutcome = {
  correct: "correct",
  incorrect: "incorrect",
  pending: "pending",
} as const;

export interface Signal {
  name: string;
  value: number;
  weight: number;
  bullish: boolean;
}

export interface Prediction {
  id: number;
  symbol: string;
  direction: PredictionDirection;
  targetPrice: number;
  currentPrice: number;
  confidence: number;
  timeframe: string;
  signals: Signal[];
  outcome?: PredictionOutcome;
  createdAt: string;
  resolvedAt?: string | null;
}

export interface CreatePredictionBody {
  symbol: string;
  timeframe?: string;
}

export interface SymbolAccuracy {
  symbol: string;
  total: number;
  correct: number;
  accuracy: number;
}

export interface PredictionSummary {
  totalPredictions: number;
  correctPredictions: number;
  accuracy: number;
  averageConfidence: number;
  recentAccuracy: number;
  improvementTrend: number;
  bySymbol: SymbolAccuracy[];
}

export interface MonteCarloBody {
  symbol: string;
  currentPrice: number;
  volatility: number;
  eventImpact: number;
  timeHorizon: number;
  simulations?: number;
}

export interface MonteCarloResult {
  symbol: string;
  simulations: number;
  median: number;
  mean: number;
  p10: number;
  p25: number;
  p75: number;
  p90: number;
  bullishProbability: number;
  bearishProbability: number;
  paths: number[][];
}

export interface PolymarketMarket {
  id: string;
  question: string;
  category?: string;
  yesPrice: number;
  noPrice: number;
  volume?: number;
  liquidity?: number;
  endDate?: string;
  active: boolean;
}

export type GetMarketHistoryParams = {
  days?: number;
};

export type GetPolymarketMarketsParams = {
  limit?: number;
};