# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hivemind Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/history/{symbol}:
    get:
      operationId: getMarketHistory
      tags: [market]
      summary: Get historical price data for a symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Historical price data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketHistory"
...
components:
  schemas:
...
    MarketHistory:
      type: object
      properties:
        symbol:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
      required: [symbol, data]
    PricePoint:
      type: object
      properties:
        timestamp:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required: [timestamp, close]
...